refactor(parsers): rename parse files to reflect their purpuse
The name of a parser should reflect what it does. A descriptive
filename has the advantage of keeping the parser single-purpose
