+++ lib/parse-imagesmin.js
