+++ lib/parse-jsmin.js
