+++ lib/parse-xmlmin.js
